requirements.txt
flask
requests
rsa
pillow
eventlet
gunicorn
git+https://github.com/Supervisor/supervisor